feat(extract): WU-2C populate real bbox from pdfjs-dist positional data
- src/extract.js: page-relative bbox percentages for matched fields (0-100, 2 decimals)
- contracts.rs: FieldBboxV1 u32→f64, drop Eq from MatchedFieldV1/InvoiceFieldsV1/LocalExtractionV1
- 6 new extract.test.js tests for bbox positioning, editable flag, null fallback
- Full suite: 216 Node + 70 Rust = 286 tests, all passing

Gotcha: pdfjs merges sibling Tj in same BT/ET block → fixtures must use separate BT/ET pairs per Tj for positional anchors.

+++ apps/nelupdf/src-tauri/src/contracts.rs
@@ -440,17 +440,22 @@ pub struct InvoiceTotalsV1 {
     pub total: Option<String>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+// Field bboxes are page-relative percentages in [0, 100]. The Node sidecar
+// rounds each value to two decimal places (see src/extract.js), so we keep
+// f64 here so the wire format can carry 4.90 / 4.29 / 1.01 etc. without
+// silently truncating to integers. f64 does not implement Eq, so the Eq
+// bound is dropped from this struct and its parents below.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(deny_unknown_fields, rename_all = "camelCase")]
 pub struct FieldBboxV1 {
     pub page: u32,
-    pub x: u32,
-    pub y: u32,
-    pub width: u32,
-    pub height: u32,
+    pub x: f64,
+    pub y: f64,
+    pub width: f64,
+    pub height: f64,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(deny_unknown_fields, rename_all = "camelCase")]
 pub struct MatchedFieldV1 {
     pub label: String,
@@ -459,7 +464,7 @@ pub struct MatchedFieldV1 {
     pub editable: bool,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(deny_unknown_fields, rename_all = "camelCase")]
 pub struct InvoiceFieldsV1 {
     pub invoice_number: Option<String>,
@@ -470,7 +475,7 @@ pub struct InvoiceFieldsV1 {
     pub matched: Vec<MatchedFieldV1>,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(deny_unknown_fields, rename_all = "camelCase")]
 pub struct LocalExtractionV1 {
     pub provenance: String,
